+++ packages/skema/src/constants.js
@@ -1,0 +1,6 @@
+import {
+  symbolFor
+} from './future'
+
+// export const IS_TYPES = symbolFor('types')
+export const IS_TYPE = symbolFor('type')
